Menu Project.
This is a project that i've just started to work on called 'Menu'. The premise of this project is to make an application that will call different sections of the program, and run/run sub-programs and on command, fall back several dirs and go back to the main menu to restart/quit program.