Cleanup/Added Menu For Modify Data Method...
[CLEANUP]
So I cleaned up some of the variables, aka in the main var section of the program, made categories of the variables for access.

[ADDED MENU FOR MODIFY DATA METHOD]
For this section, if the user wants to modify the data, now they get a menu on who they want to edit.

[FOR NEXT VERSION]
Add onto the modify data method, where they select what they want to change about the person.