[AUTH] Check around program to see when user can add more family members.
This commit is a bit confusing so let me explain.

The booleans are the heart of the commit, they allow data to be added via the menu. Once a user adds family data, the boolean is true, as now the family member has data attached to their name. If the user then decides to delete the family members data, then the boolean becomes false, and because it's false, it can be used in the add family members tab.